requirements.txt
pydicom>=2.4.4
Pillow>=10.0.0
numpy>=1.26.4
pandas>=2.2.2
tqdm>=4.66.4
openai>=1.14.0
reportlab>=4.1.0
httpx>=0.27.0
pytest>=8.0.0
python-dotenv>=1.0.0 